Better maintain correct project structure for routing
